Rename old transformer.py -> transformer_config.py
PiperOrigin-RevId: 753100992
