+++ gemma/gm/math/_pos_utils.py
@@ -1,0 +1,35 @@
+# Copyright 2024 DeepMind Technologies Limited.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Utils for computing positions."""
+
+import jax
+import jax.numpy as jnp
+
+
+# TODO(epot): Add real typing array annotations.
+def build_positions_from_mask(input_mask: jax.Array) -> jax.Array:
+  """Computes the `positions` from the `input_mask`.
+
+  Args:
+    input_mask: The tokens `input_mask`, True for non-padded tokens only.
+
+  Returns:
+    The indices to use for RoPE and absolute position encodings for the given
+    input mask.
+  """
+  positions = jnp.cumsum(input_mask, axis=-1)
+  # Subtract one for all positions from the first valid one as they are
+  # 0-indexed
+  return positions - (positions >= 1)
