+++ gemma/gm/nn/_config.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Gemma transformer."""
+"""Gemma transformer config."""
 
 from collections.abc import Sequence
 import dataclasses
@@ -21,7 +21,6 @@ import functools
 
 from gemma import modules
 from gemma.multimodal import vision as gemma_vision
-import jax
 import jax.numpy as jnp
 
 Cache = dict[str, modules.LayerCache]
@@ -119,19 +118,3 @@ class TransformerConfig:
         for i in range(self.num_layers)
     }
     return cache
-
-
-def build_positions_from_mask(input_mask: jax.Array) -> jax.Array:
-  """Computes the `positions` from the `input_mask`.
-
-  Args:
-    input_mask: The tokens `input_mask`, True for non-padded tokens only.
-
-  Returns:
-    The indices to use for RoPE and absolute position encodings for the given
-    input mask.
-  """
-  positions = jnp.cumsum(input_mask, axis=-1)
-  # Subtract one for all positions from the first valid one as they are
-  # 0-indexed
-  return positions - (positions >= 1)
