+++ gemma/gm/nn/config.py
@@ -19,9 +19,9 @@
 from etils import epy as _epy
 
 with _epy.lazy_api_imports(globals()):
+  from gemma.gm.nn._config import make_attention_layers_types
+  from gemma.gm.nn._config import QueryPreAttentionNormalisation
+  from gemma.gm.nn._config import TransformerConfig
   from gemma.gm.nn._gemma import GEMMA3_ATTENTION_PATTERN
   from gemma.gm.nn._transformer import ModelInfo
   from gemma.modules import AttentionType
-  from gemma.transformer import make_attention_layers_types
-  from gemma.transformer import QueryPreAttentionNormalisation
-  from gemma.transformer import TransformerConfig
